moving notifications page to account folder
