+++ fresh.gen.ts
@@ -9,23 +9,22 @@ import * as $3 from "./routes/_middleware.ts";
 import * as $4 from "./routes/account/_middleware.ts";
 import * as $5 from "./routes/account/index.tsx";
 import * as $6 from "./routes/account/manage.ts";
-import * as $7 from "./routes/account/upgrade.ts";
-import * as $8 from "./routes/api/stripe-webhooks.ts";
-import * as $9 from "./routes/api/vote.ts";
-import * as $10 from "./routes/blog/[slug].tsx";
-import * as $11 from "./routes/blog/index.tsx";
-import * as $12 from "./routes/callback.ts";
-import * as $13 from "./routes/feed.ts";
-import * as $14 from "./routes/index.tsx";
-import * as $15 from "./routes/item/[id].tsx";
-import * as $16 from "./routes/notifications/_middleware.ts";
-import * as $17 from "./routes/notifications/index.tsx";
-import * as $18 from "./routes/pricing.tsx";
-import * as $19 from "./routes/signin.ts";
-import * as $20 from "./routes/signout.ts";
-import * as $21 from "./routes/stats.tsx";
-import * as $22 from "./routes/submit.tsx";
-import * as $23 from "./routes/user/[username].tsx";
+import * as $7 from "./routes/account/notifications.tsx";
+import * as $8 from "./routes/account/upgrade.ts";
+import * as $9 from "./routes/api/stripe-webhooks.ts";
+import * as $10 from "./routes/api/vote.ts";
+import * as $11 from "./routes/blog/[slug].tsx";
+import * as $12 from "./routes/blog/index.tsx";
+import * as $13 from "./routes/callback.ts";
+import * as $14 from "./routes/feed.ts";
+import * as $15 from "./routes/index.tsx";
+import * as $16 from "./routes/item/[id].tsx";
+import * as $17 from "./routes/pricing.tsx";
+import * as $18 from "./routes/signin.ts";
+import * as $19 from "./routes/signout.ts";
+import * as $20 from "./routes/stats.tsx";
+import * as $21 from "./routes/submit.tsx";
+import * as $22 from "./routes/user/[username].tsx";
 import * as $$0 from "./islands/PageInput.tsx";
 import * as $$1 from "./islands/VoteButton.tsx";
 
@@ -38,23 +37,22 @@ const manifest = {
     "./routes/account/_middleware.ts": $4,
     "./routes/account/index.tsx": $5,
     "./routes/account/manage.ts": $6,
-    "./routes/account/upgrade.ts": $7,
-    "./routes/api/stripe-webhooks.ts": $8,
-    "./routes/api/vote.ts": $9,
-    "./routes/blog/[slug].tsx": $10,
-    "./routes/blog/index.tsx": $11,
-    "./routes/callback.ts": $12,
-    "./routes/feed.ts": $13,
-    "./routes/index.tsx": $14,
-    "./routes/item/[id].tsx": $15,
-    "./routes/notifications/_middleware.ts": $16,
-    "./routes/notifications/index.tsx": $17,
-    "./routes/pricing.tsx": $18,
-    "./routes/signin.ts": $19,
-    "./routes/signout.ts": $20,
-    "./routes/stats.tsx": $21,
-    "./routes/submit.tsx": $22,
-    "./routes/user/[username].tsx": $23,
+    "./routes/account/notifications.tsx": $7,
+    "./routes/account/upgrade.ts": $8,
+    "./routes/api/stripe-webhooks.ts": $9,
+    "./routes/api/vote.ts": $10,
+    "./routes/blog/[slug].tsx": $11,
+    "./routes/blog/index.tsx": $12,
+    "./routes/callback.ts": $13,
+    "./routes/feed.ts": $14,
+    "./routes/index.tsx": $15,
+    "./routes/item/[id].tsx": $16,
+    "./routes/pricing.tsx": $17,
+    "./routes/signin.ts": $18,
+    "./routes/signout.ts": $19,
+    "./routes/stats.tsx": $20,
+    "./routes/submit.tsx": $21,
+    "./routes/user/[username].tsx": $22,
   },
   islands: {
     "./islands/PageInput.tsx": $$0,
